+++ src/info/authors.rs
@@ -1,10 +1,10 @@
-use super::git::metrics::GitMetrics;
+use super::git::sig::Sig;
 use crate::{
     cli::NumberSeparator,
     info::utils::{format_number, info_field::InfoField},
 };
 use serde::Serialize;
-use std::fmt::Write;
+use std::{collections::HashMap, fmt::Write};
 
 #[derive(Serialize, Clone, Debug, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -66,8 +66,20 @@ pub struct AuthorsInfo {
 }
 
 impl AuthorsInfo {
-    pub fn new(git_metrics: &GitMetrics) -> Self {
-        let authors = git_metrics.authors_to_display.clone();
+    pub fn new(
+        number_of_commits_by_signature: &HashMap<Sig, usize>,
+        total_number_of_commits: usize,
+        number_of_authors_to_display: usize,
+        show_email: bool,
+        number_separator: NumberSeparator,
+    ) -> Self {
+        let authors = compute_authors(
+            number_of_commits_by_signature,
+            total_number_of_commits,
+            number_of_authors_to_display,
+            show_email,
+            number_separator,
+        );
         Self { authors }
     }
 
@@ -77,6 +89,40 @@ impl AuthorsInfo {
         }
         0
     }
+}
+
+fn compute_authors(
+    number_of_commits_by_signature: &HashMap<Sig, usize>,
+    total_number_of_commits: usize,
+    number_of_authors_to_display: usize,
+    show_email: bool,
+    number_separator: NumberSeparator,
+) -> Vec<Author> {
+    let mut signature_with_number_of_commits_sorted: Vec<(&Sig, &usize)> =
+        Vec::from_iter(number_of_commits_by_signature);
+
+    signature_with_number_of_commits_sorted.sort_by(|(sa, a_count), (sb, b_count)| {
+        b_count.cmp(a_count).then_with(|| sa.name.cmp(&sb.name))
+    });
+
+    let authors: Vec<Author> = signature_with_number_of_commits_sorted
+        .into_iter()
+        .map(|(author, author_nbr_of_commits)| {
+            Author::new(
+                author.name.to_string(),
+                if show_email {
+                    Some(author.email.to_string())
+                } else {
+                    None
+                },
+                *author_nbr_of_commits,
+                total_number_of_commits,
+                number_separator,
+            )
+        })
+        .take(number_of_authors_to_display)
+        .collect();
+    authors
 }
 
 fn digit_difference(num1: usize, num2: usize) -> usize {
@@ -284,5 +330,49 @@ mod test {
     fn test_digit_difference(#[case] num1: usize, #[case] num2: usize, #[case] expected: usize) {
         let result = digit_difference(num1, num2);
         assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn test_compute_authors() {
+        let mut number_of_commits_by_signature: HashMap<Sig, usize> = HashMap::new();
+        number_of_commits_by_signature.insert(
+            Sig {
+                name: "John Doe".into(),
+                email: "johndoe@example.com".into(),
+            },
+            30,
+        );
+        number_of_commits_by_signature.insert(
+            Sig {
+                name: "Jane Doe".into(),
+                email: "janedoe@example.com".into(),
+            },
+            20,
+        );
+        number_of_commits_by_signature.insert(
+            Sig {
+                name: "Ellen Smith".into(),
+                email: "ellensmith@example.com".into(),
+            },
+            50,
+        );
+        let total_number_of_commits = 100;
+        let number_of_authors_to_display = 2;
+        let show_email = false;
+        let number_separator = NumberSeparator::Comma;
+
+        let actual = compute_authors(
+            &number_of_commits_by_signature,
+            total_number_of_commits,
+            number_of_authors_to_display,
+            show_email,
+            number_separator,
+        );
+
+        let expected = vec![
+            Author::new(String::from("Ellen Smith"), None, 50, 100, number_separator),
+            Author::new(String::from("John Doe"), None, 30, 100, number_separator),
+        ];
+        assert_eq!(actual, expected);
     }
 }
